data: cleanup - comments and move field to own file also sql test to own file
