+++ data/field_config.go
@@ -49,6 +49,7 @@ type FieldConfig struct {
 // to null.
 type ConfFloat64 float64
 
+// MarshalJSON fullfills the json.Marshaler interface.
 func (sf *ConfFloat64) MarshalJSON() ([]byte, error) {
 	if sf == nil || math.IsNaN(float64(*sf)) || math.IsInf(float64(*sf), -1) || math.IsInf(float64(*sf), 1) {
 		return []byte("null"), nil
@@ -57,6 +58,7 @@ func (sf *ConfFloat64) MarshalJSON() ([]byte, error) {
 	return []byte(fmt.Sprintf(`%v`, float64(*sf))), nil
 }
 
+// UnmarshalJSON fullfills the json.Unmarshaler interface.
 func (sf *ConfFloat64) UnmarshalJSON(data []byte) error {
 	s := string(data)
 	if s == "null" {
