+++ data/frame_test.go
@@ -1,10 +1,7 @@
 package data_test
 
 import (
-	"database/sql"
 	"fmt"
-	"reflect"
-	"strconv"
 	"testing"
 	"time"
 
@@ -35,87 +32,6 @@ func TestFrameWarnings(t *testing.T) {
 
 	if len(df.Warnings) != 2 {
 		t.Fatal("expected two warnings to be appended")
-	}
-}
-
-func TestField(t *testing.T) {
-	f := data.NewField("value", nil, []float64{1.0, 2.0, 3.0})
-
-	if f.Len() != 3 {
-		t.Fatal("unexpected length")
-	}
-}
-
-func TestField_Float64(t *testing.T) {
-	field := data.NewField("value", nil, make([]*float64, 3))
-
-	want := 2.0
-	field.Set(1, &want)
-
-	if field.Len() != 3 {
-		t.Fatal("unexpected length")
-	}
-
-	got := field.At(1).(*float64)
-
-	if *got != want {
-		t.Errorf("%+v", *got)
-	}
-}
-
-func TestField_String(t *testing.T) {
-	field := data.NewField("value", nil, make([]*string, 3))
-
-	want := "foo"
-	field.Set(1, &want)
-
-	if field.Len() != 3 {
-		t.Fatal("unexpected length")
-	}
-
-	got := field.At(1).(*string)
-
-	if *got != want {
-		t.Errorf("%+v", *got)
-	}
-}
-
-func TestTimeField(t *testing.T) {
-	tests := []struct {
-		Values []*time.Time
-	}{
-		{
-			Values: []*time.Time{timePtr(time.Unix(111, 0))},
-		},
-		{
-			Values: []*time.Time{nil, timePtr(time.Unix(111, 0))},
-		},
-		{
-			Values: []*time.Time{nil, timePtr(time.Unix(111, 0)), nil},
-		},
-		{
-			Values: make([]*time.Time, 10),
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run("", func(t *testing.T) {
-			f := data.NewField(t.Name(), nil, tt.Values)
-
-			if f.Len() != len(tt.Values) {
-				t.Error(f.Len())
-			}
-
-			for i := 0; i < f.Len(); i++ {
-				got := reflect.ValueOf(f.At(i))
-				want := reflect.ValueOf(tt.Values[i])
-
-				if got != want {
-					t.Error(got, want)
-				}
-			}
-
-		})
 	}
 }
 
@@ -328,63 +244,4 @@ func stringPtr(s string) *string {
 
 func boolPtr(b bool) *bool {
 	return &b
-}
-
-func ExampleSQLStringConverter() {
-	_ = data.SQLStringConverter{
-		Name:          "BIGINT to *int64",
-		InputScanKind: reflect.Struct,
-		InputTypeName: "BIGINT",
-		Replacer: &data.StringFieldReplacer{
-			VectorType: []*int64{},
-			ReplaceFunc: func(in *string) (interface{}, error) {
-				if in == nil {
-					return nil, nil
-				}
-				v, err := strconv.ParseInt(*in, 10, 64)
-				if err != nil {
-					return nil, err
-				}
-				return &v, nil
-			},
-		},
-	}
-}
-
-func ExampleStringFieldReplacer() {
-	_ = &data.StringFieldReplacer{
-		VectorType: []*int64{},
-		ReplaceFunc: func(in *string) (interface{}, error) {
-			if in == nil {
-				return nil, nil
-			}
-			v, err := strconv.ParseInt(*in, 10, 64)
-			if err != nil {
-				return nil, err
-			}
-			return &v, nil
-		},
-	}
-}
-
-func ExampleNewFromSQLRows() {
-	aQuery := "SELECT * FROM GoodData"
-	db, err := sql.Open("fancySql", "fancysql://user:pass@localhost:1433")
-	if err != nil {
-		// return err
-	}
-
-	defer db.Close()
-
-	rows, err := db.Query(aQuery)
-	if err != nil {
-		// return err
-	}
-	defer rows.Close()
-
-	frame, mappings, err := data.NewFromSQLRows(rows)
-	if err != nil {
-		// return err
-	}
-	_, _ = frame, mappings
 }
